Update packages and template

+++ dist.sh
@@ -1,13 +1,23 @@
 #!/usr/bin/env bash
 
 if [[ $# -eq 0 ]] ; then
-    echo 'Missing APP_ENV parameter'
+    echo 'Missing NODE_ENV, APP_ENV, and APP_TYPE parameters'
+    exit 0
+fi
+
+if [[ $# -eq 1 ]] ; then
+    echo 'Missing APP_ENV, and APP_TYPE parameters'
+    exit 0
+fi
+
+if [[ $# -eq 2 ]] ; then
+    echo 'Missing APP_TYPE parameter'
     exit 0
 fi
 
 rm -rf ./build
 rm -rf ./dist
 mkdir dist
-NODE_ENV=$1 APP_ENV=$1 ./node_modules/webpack/bin/webpack.js -p
+NODE_ENV=$1 APP_ENV=$2 APP_TYPE=$3 ./node_modules/webpack/bin/webpack.js
 
-zip -r dist/build-frontend-$1 build
+zip -r dist/build-frontend-$3 build
